+++ code/Graphics/Mesh/Mesh.hpp
@@ -14,13 +14,13 @@ public:
     void addNormal(const Vector3& vector);
     void addTexCoord(f32 u, f32 v);
     void addColor(f32 r, f32 g, f32 b, f32 a);
-    void addFace(u16 v1, u16 v2, u16 v3);
-    void addFaceIndex(u16 i);
+    void addFace(u32 v1, u32 v2, u32 v3);
+    void addFaceIndex(u32 i);
     void addVertices(const std::vector<f32> vec);
     void addNormals(const std::vector<f32> vec);
     void addTextureCoordinates(const std::vector<f32> vec);
     void addColors(const std::vector<f32> vec);
-    void addFaces(const std::vector<u16> vec);
+    void addFaces(const std::vector<u32> vec);
     void copyVertices(Ptr<const Mesh> other);
     void copyNormals(Ptr<const Mesh> other);
     void copyTextureCoordinates(Ptr<const Mesh> other);
@@ -33,7 +33,7 @@ private:
 	std::vector<f32> mNormals;
 	std::vector<f32> mTextureCoordinates;
 	std::vector<f32> mColors;
-	std::vector<u16> mFaces;
+	std::vector<u32> mFaces;
 
 	u32 mVertexCount = 0;
 	u32 mFacesCount = 0;
